pyproject.toml
[project]
name = "nanokvm-mcp"
version = "0.1.0"
description = "MCP server for controlling NanoKVM devices"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "pycryptodome>=3.20.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require mocked API)",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["nanokvm_mcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[project.scripts]
nanokvm-mcp = "nanokvm_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanokvm_mcp"]
